Move killall to sglang.cli, add CI/local dual mode, add .sh backward compat
- Move killall_sglang.py to python/sglang/cli/killall.py
- Add dual mode: CI mode (SGLANG_IS_IN_CI) uses GPU-scoped kill via
  nvidia-smi; local mode uses pgrep-based kill (original .sh behavior)
- Warn if CUDA_VISIBLE_DEVICES not set in CI mode
- Add killall_sglang console_script entry point in pyproject.toml
- Add scripts/killall_sglang.sh as deprecated backward compat wrapper
- Update CI call sites to new path
